chore: fix cicd errors
